+++ mod-data-import-converter-storage-server/src/test/java/org/folio/rest/impl/MappingProfileTest.java
@@ -9,7 +9,7 @@ import io.vertx.ext.unit.junit.VertxUnitRunner;
 import org.apache.http.HttpStatus;
 import org.folio.rest.jaxrs.model.ActionProfile;
 import org.folio.rest.jaxrs.model.MappingProfile;
-import org.folio.rest.jaxrs.model.MatchProfile;
+import org.folio.rest.jaxrs.model.MappingProfile.IncomingRecordType;
 import org.folio.rest.jaxrs.model.ProfileAssociation;
 import org.folio.rest.jaxrs.model.Tags;
 import org.folio.rest.persist.Criteria.Criterion;
@@ -27,6 +27,7 @@ import static org.folio.rest.impl.ActionProfileTest.ACTION_PROFILES_PATH;
 import static org.folio.rest.impl.ActionProfileTest.ACTION_PROFILES_TABLE_NAME;
 import static org.folio.rest.jaxrs.model.ActionProfile.Action.CREATE;
 import static org.folio.rest.jaxrs.model.ActionProfile.FolioRecord.MARC_BIBLIOGRAPHIC;
+import static org.folio.rest.jaxrs.model.MappingProfile.FolioRecord.INSTANCE;
 import static org.folio.rest.jaxrs.model.ProfileSnapshotWrapper.ContentType.ACTION_PROFILE;
 import static org.folio.rest.jaxrs.model.ProfileSnapshotWrapper.ContentType.MAPPING_PROFILE;
 import static org.hamcrest.Matchers.empty;
@@ -43,11 +44,17 @@ public class MappingProfileTest extends AbstractRestVerticleTest {
   private static final String ASSOCIATED_PROFILES_PATH = "/data-import-profiles/profileAssociations";
 
   private static MappingProfile mappingProfile_1 = new MappingProfile().withName("Bla")
-    .withTags(new Tags().withTagList(Arrays.asList("lorem", "ipsum", "dolor")));
+    .withTags(new Tags().withTagList(Arrays.asList("lorem", "ipsum", "dolor")))
+    .withIncomingRecordType(IncomingRecordType.MARC_BIBLIOGRAPHIC)
+    .withFolioRecord(INSTANCE);
   private static MappingProfile mappingProfile_2 = new MappingProfile().withName("Boo")
-    .withTags(new Tags().withTagList(Arrays.asList("lorem", "ipsum")));
+    .withTags(new Tags().withTagList(Arrays.asList("lorem", "ipsum")))
+    .withIncomingRecordType(IncomingRecordType.MARC_BIBLIOGRAPHIC)
+    .withFolioRecord(INSTANCE);
   private static MappingProfile mappingProfile_3 = new MappingProfile().withName("Foo")
-    .withTags(new Tags().withTagList(Collections.singletonList("lorem")));
+    .withTags(new Tags().withTagList(Collections.singletonList("lorem")))
+    .withIncomingRecordType(IncomingRecordType.MARC_BIBLIOGRAPHIC)
+    .withFolioRecord(INSTANCE);
 
   @Test
   public void shouldReturnEmptyListOnGet() {
@@ -178,7 +185,10 @@ public class MappingProfileTest extends AbstractRestVerticleTest {
 
     Response createResponse = RestAssured.given()
       .spec(spec)
-      .body(new MappingProfile().withName("newProfile"))
+      .body(new MappingProfile()
+        .withName("newProfile")
+        .withIncomingRecordType(IncomingRecordType.MARC_BIBLIOGRAPHIC)
+        .withFolioRecord(INSTANCE))
       .when()
       .post(MAPPING_PROFILES_PATH);
     Assert.assertThat(createResponse.statusCode(), is(HttpStatus.SC_CREATED));
@@ -202,7 +212,7 @@ public class MappingProfileTest extends AbstractRestVerticleTest {
       .when()
       .post(MAPPING_PROFILES_PATH);
     Assert.assertThat(createResponse.statusCode(), is(HttpStatus.SC_CREATED));
-    MatchProfile mappingProfile = createResponse.body().as(MatchProfile.class);
+    MappingProfile mappingProfile = createResponse.body().as(MappingProfile.class);
 
     mappingProfile.setDescription("test");
     RestAssured.given()
@@ -239,7 +249,7 @@ public class MappingProfileTest extends AbstractRestVerticleTest {
       .when()
       .post(MAPPING_PROFILES_PATH);
     Assert.assertThat(createResponse.statusCode(), is(HttpStatus.SC_CREATED));
-    MatchProfile mappingProfile = createResponse.body().as(MatchProfile.class);
+    MappingProfile mappingProfile = createResponse.body().as(MappingProfile.class);
 
     RestAssured.given()
       .spec(spec)
@@ -316,7 +326,7 @@ public class MappingProfileTest extends AbstractRestVerticleTest {
       .when()
       .post(MAPPING_PROFILES_PATH);
     Assert.assertThat(createResponse.statusCode(), is(HttpStatus.SC_CREATED));
-    MatchProfile profile = createResponse.body().as(MatchProfile.class);
+    MappingProfile profile = createResponse.body().as(MappingProfile.class);
 
     RestAssured.given()
       .spec(spec)
@@ -340,7 +350,9 @@ public class MappingProfileTest extends AbstractRestVerticleTest {
     MappingProfile mappingProfileToDelete = RestAssured.given()
       .spec(spec)
       .body(new MappingProfile()
-        .withName("ProfileToDelete"))
+        .withName("ProfileToDelete")
+        .withIncomingRecordType(IncomingRecordType.MARC_BIBLIOGRAPHIC)
+        .withFolioRecord(INSTANCE))
       .when()
       .post(MAPPING_PROFILES_PATH)
       .then()
@@ -370,7 +382,9 @@ public class MappingProfileTest extends AbstractRestVerticleTest {
     MappingProfile mappingProfileToDelete = RestAssured.given()
       .spec(spec)
       .body(new MappingProfile()
-        .withName("ProfileToDelete"))
+        .withName("ProfileToDelete")
+        .withIncomingRecordType(IncomingRecordType.MARC_BIBLIOGRAPHIC)
+        .withFolioRecord(INSTANCE))
       .when()
       .post(MAPPING_PROFILES_PATH)
       .then()