+++ mod-data-import-converter-storage-server/src/test/java/org/folio/rest/impl/association/CommonProfileAssociationTest.java
@@ -31,6 +31,7 @@ import java.util.UUID;
 import static org.folio.rest.jaxrs.model.ActionProfile.Action.CREATE;
 import static org.folio.rest.jaxrs.model.ActionProfile.FolioRecord.MARC_BIBLIOGRAPHIC;
 import static org.folio.rest.jaxrs.model.JobProfile.DataType.MARC;
+import static org.folio.rest.jaxrs.model.MappingProfile.FolioRecord.INSTANCE;
 import static org.folio.rest.jaxrs.model.ProfileSnapshotWrapper.ContentType.ACTION_PROFILE;
 import static org.folio.rest.jaxrs.model.ProfileSnapshotWrapper.ContentType.JOB_PROFILE;
 import static org.folio.rest.jaxrs.model.ProfileSnapshotWrapper.ContentType.MAPPING_PROFILE;
@@ -75,9 +76,15 @@ public class CommonProfileAssociationTest extends AbstractRestVerticleTest {
   ActionProfile actionProfile4 = new ActionProfile().withName("testActionProfile4")
     .withAction(CREATE).withFolioRecord(MARC_BIBLIOGRAPHIC);
 
-  MappingProfile mappingProfile1 = new MappingProfile().withName("testMappingProfile1").withDescription("test-description");
-  MappingProfile mappingProfile2 = new MappingProfile().withName("testMappingProfile2").withDescription("test-description");
-  MappingProfile mappingProfile3 = new MappingProfile().withName("testMappingProfile3");
+  MappingProfile mappingProfile1 = new MappingProfile().withName("testMappingProfile1").withDescription("test-description")
+    .withIncomingRecordType(MappingProfile.IncomingRecordType.MARC_BIBLIOGRAPHIC)
+    .withFolioRecord(INSTANCE);
+  MappingProfile mappingProfile2 = new MappingProfile().withName("testMappingProfile2").withDescription("test-description")
+    .withIncomingRecordType(MappingProfile.IncomingRecordType.MARC_BIBLIOGRAPHIC)
+    .withFolioRecord(INSTANCE);
+  MappingProfile mappingProfile3 = new MappingProfile().withName("testMappingProfile3")
+    .withIncomingRecordType(MappingProfile.IncomingRecordType.MARC_BIBLIOGRAPHIC)
+    .withFolioRecord(INSTANCE);
 
   MatchProfile matchProfile1 = new MatchProfile().withName("testMatchProfile1")
     .withIncomingRecordType(MatchProfile.IncomingRecordType.MARC)